Fix codegen register allocation and truncation bugs
- Fix compile_destructuring: was freeing val_reg before declaring as
  local, allowing subsequent allocations to clobber it. Now frees both
  temp registers and re-allocates for the local.
- Fix emit_jump: change uint8_t a parameter to int to avoid truncating
  register indices > 255
- Fix CALL/TAIL_CALL arg count: remove static_cast<uint8_t> that
  truncated argument counts > 255

+++ include/akar/compiler/codegen.h
@@ -103,7 +103,7 @@ private:
     void emit(uint8_t op, int a, int b, int c);
     void emit_bx(uint8_t op, int a, uint16_t bx);
     void emit_wide(uint8_t op, int a, int b, int c);
-    size_t emit_jump(uint8_t op, uint8_t a, int16_t offset = 0);
+    size_t emit_jump(uint8_t op, int a, int16_t offset = 0);
     void patch_jump(size_t jump_byte_pos, int16_t offset);
     size_t current_offset() const;
     size_t add_constant(Value value);
